feat: add passwordValidationRegex property
