+++ packages/api/src/v1/profile.ts
@@ -2,8 +2,15 @@ import { Response, Router } from 'express';
 import _ from 'lodash';
 
 import { IAuth } from '@verdaccio/auth';
-import { API_ERROR, APP_ERROR, HTTP_STATUS, SUPPORT_ERRORS, errorUtils } from '@verdaccio/core';
-import { validatePassword } from '@verdaccio/utils';
+import {
+  API_ERROR,
+  APP_ERROR,
+  HTTP_STATUS,
+  SUPPORT_ERRORS,
+  errorUtils,
+  validatioUtils,
+} from '@verdaccio/core';
+import { Config } from '@verdaccio/types';
 
 import { $NextFunctionVer, $RequestExtend } from '../../types/custom';
 
@@ -18,7 +25,7 @@ export interface Profile {
   fullname: string;
 }
 
-export default function (route: Router, auth: IAuth): void {
+export default function (route: Router, auth: IAuth, config: Config): void {
   function buildProfile(name: string): Profile {
     return {
       tfa: false,
@@ -60,9 +67,14 @@ export default function (route: Router, auth: IAuth): void {
       const { name } = req.remote_user;
 
       if (_.isNil(password) === false) {
-        if (validatePassword(password.new) === false) {
+        if (
+          validatioUtils.validatePassword(
+            password.new,
+            config?.serverSettings?.passwordValidationRegex
+          ) === false
+        ) {
           /* eslint new-cap:off */
-          return next(errorUtils.getCode(HTTP_STATUS.UNAUTHORIZED, API_ERROR.PASSWORD_SHORT()));
+          return next(errorUtils.getCode(HTTP_STATUS.UNAUTHORIZED, API_ERROR.PASSWORD_SHORT));
           /* eslint new-cap:off */
         }
