+++ packages/auth/src/auth.ts
@@ -142,7 +142,7 @@ class Auth implements IAuth {
         // @ts-ignore
         return authenticate || allow_access || allow_publish;
       },
-      this.config?.server?.pluginPrefix
+      this.config?.serverSettings?.pluginPrefix
     );
   }
