+++ packages/core/core/src/error-utils.ts
@@ -1,11 +1,9 @@
 import createError, { HttpError } from 'http-errors';
 
-import { DEFAULT_MIN_LIMIT_PASSWORD, HTTP_STATUS } from './constants';
+import { HTTP_STATUS } from './constants';
 
 export const API_ERROR = {
-  PASSWORD_SHORT: (passLength = DEFAULT_MIN_LIMIT_PASSWORD): string =>
-    `The provided password is too short. Please pick a password longer than ` +
-    `${passLength} characters.`,
+  PASSWORD_SHORT: `The provided password does not pass the validation`,
   MUST_BE_LOGGED: 'You must be logged in to publish packages.',
   PLUGIN_ERROR: 'bug in the auth plugin system',
   CONFIG_BAD_FORMAT: 'config file must be an object',
