+++ packages/core/types/src/configuration.ts
@@ -233,6 +233,7 @@ export type ServerSettingsConf = {
    * acme-XXXXXX
    */
   pluginPrefix?: string;
+  passwordValidationRegex?: RegExp;
 };
 
 /**
