+++ packages/utils/test/auth-utils.spec.ts
@@ -1,24 +1,6 @@
-import { createSessionToken, getAuthenticatedMessage, validatePassword } from '../src';
+import { createSessionToken, getAuthenticatedMessage } from '../src';
 
 describe('Auth Utilities', () => {
-  describe('validatePassword', () => {
-    test('should validate password according the length', () => {
-      expect(validatePassword('12345', 1)).toBeTruthy();
-    });
-
-    test('should fails on validate password according the length', () => {
-      expect(validatePassword('12345', 10)).toBeFalsy();
-    });
-
-    test('should fails on validate password according the length and default config', () => {
-      expect(validatePassword('12')).toBeFalsy();
-    });
-
-    test('should validate password according the length and default config', () => {
-      expect(validatePassword('1235678910')).toBeTruthy();
-    });
-  });
-
   describe('createSessionToken', () => {
     test('should generate session token', () => {
       expect(createSessionToken()).toHaveProperty('expires');
